openAI_API/requirements.txt
langgraph>=0.2.50
langgraph-checkpoint-sqlite>=2.0
langchain-openai>=0.3
langchain-core>=0.3
pydantic>=2.0
httpx>=0.27
trafilatura>=1.12
python-dotenv>=1.0
langchain-text-splitters>=0.3
lancedb>=0.10
pyarrow
rich>=13.0
